Honor the interval parameter
This change will use the --interval (-i) parameter supplied on the
command line to delay the execution of the next metrics retrieval.

Previously, the interval was hard-coded for 1 minute (60 secs).

It's probably not a good idea to set the interval to less than
10 seconds. Retrieving metrics will take about this long and
graphite has a default resolution of 10 seconds.